+++ Libraries/KLibC/stdint/types/uint.h
@@ -2,10 +2,10 @@
 
 #include <assert.h>
 
-typedef unsigned char      uint8_t;
-typedef unsigned short     uint16_t;
-typedef unsigned int       uint32_t;
-typedef unsigned long long uint64_t;
+typedef unsigned char  uint8_t;
+typedef unsigned short uint16_t;
+typedef unsigned int   uint32_t;
+typedef unsigned long  uint64_t;
 
 typedef uint64_t uintmax_t;
 typedef uint64_t uintptr_t;
